GetXRButton
Allows you to pick a controller, button, and button action (buttonup, buttondown) to trigger the next action with the choice of saving the result to an FSM Variable. Utilizes the existing unity events in the package. I was used to this functionality through my last package, wanted to replicate it.